cloud_functions/slack_messages_to_bigquery/requirements.txt
google-api-python-client==2.92.0
google-auth==2.23.2
google-cloud-storage==2.11.0
Flask==2.3.3
functions-framework==3.*
google-cloud-aiplatform==1.60.0
google-cloud-secret-manager==2.16.2
google-cloud-bigquery==3.25.0
pandas==2.2.2
requests==2.31.0
pyarrow==17.0.0
google-cloud-logging==3.11.2